Create a Dockerfile to build the website output.
This also updates the Makefile to use the more modern sassc command line
tool, since the old Ruby-based scss is no longer supported.

Note that the syntax is _almost_ but not quite the same between the two
tools.

Also build the image at push time so that it can be fetched. Note that
the image does not use the "published" branch, so the link to Amazon is
not going to be present.